requirements.txt
Cython==0.22.1
google==1.7
ipython==3.2.0
numpy==1.9.2
Pillow==2.9.0
protobuf==2.6.1
scikit-image==0.11.3
scipy==0.15.1